Give up LESS and use CDN for Bootstrap and ACE
ACE editor now uses a AWS cloudfront service. Not sure if it's
gonna hold though, will check when we get to the editor part.

I don't like the glyphicon lib Bootstrap 3 uses. The FontAwesome
folks seem to be ready, just waiting for Bootstrap 3.0 official
release. I'll just wait for that, and before then let's just
ignore the icons and get the functionality done.