Split users and blogs and implement blog edit view
URLs in users app are splitted into users/ and account/ paths,
the former contains public user pages; the latter contain user
profile editing pages, which is "current-user-only."

The blog list view is implemented by adding `blog_slug` value into
the context. All views in the app MUST contain a `slug` variable
in the kwargs -- which is okay because they need it anyway.